pyproject.toml
[tool.poetry]
name = "sp_score"
version = "0.1.0"
description = "\"A score to find polymer synthesizability\""
authors = ["arajan63 <arajan63@gatech.edu>"]
license = "Ramprasad Group"

[tool.poetry.dependencies]
python = "^3.8"
rdkit-pypi = "^2021.3.4"
pytest = "^6.2.4"
pandas = "^1.3.1"
networkx = "^2.6.2"
pathos = "^0.2.8"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
